Maker rendering does not get right position.
As long as the functions AdjustForViewBoxWidth and
AdjustForViewBoxHeight are not fully developed pay attention that the
ViewBox width and height will not mismatch the positions of the marker
element then one of them are zero.

This changes are based on Issue #212 from mater.

Add: Quick solution to avoid  ViewBox width or height zero calculation.
Add: UnitTests