Found the breaking css line and simplified tests
Found the exact line in the CSS that failed (border with !important) and isolated the test to test on this. Also updated names of the tests and files. Added a reference test for the non important line (which should always work).

Also fixed small spelling error in the css parser.